requirements.txt
Django~=3.2
psycopg2-binary
djangorestframework~=3.12
markdown~=3.3.6
django-filter~=21.1
httpie~=3.0.0
djangorestframework-simplejwt~=5.1.0
